test(adapters): gate network-dependent syft tests behind integration tag
Split Test_syftAdapter_CreateSBOM, which pulls real images from public
registries, into a separate file built only with -tags=integration.
`go test ./adapters/v1` now stays hermetic and fast by default.
